feat: more flexible instance view spec
Use a specification for instance views so they can be composed from instance elements instead of there being a limited set of fixed instance views.
