+++ backend/zeno_backend/instance_views/chatbot-markdown.json
@@ -1,0 +1,22 @@
+{
+  "data": {
+    "type": "list",
+    "elements": {
+      "type": "message",
+      "content": {
+        "type": "markdown"
+      }
+    },
+    "collapsible": "top"
+  },
+  "label": { "type": "markdown" },
+  "output": {
+    "type": "message",
+    "plain": true,
+    "role": "assistant",
+    "highlight": true,
+    "content": {
+      "type": "markdown"
+    }
+  }
+}
