+++ backend/zeno_backend/instance_views/code-generation.json
@@ -1,0 +1,5 @@
+{
+  "data": { "type": "code" },
+  "label": { "type": "code" },
+  "output": { "type": "code" }
+}
