+++ backend/zeno_backend/instance_views/image-classification.json
@@ -1,0 +1,5 @@
+{
+  "data": { "type": "image" },
+  "label": { "type": "text" },
+  "output": { "type": "text" }
+}
