+++ backend/zeno_backend/instance_views/text-classification.json
@@ -1,0 +1,5 @@
+{
+  "data": { "type": "text" },
+  "label": { "type": "text" },
+  "output": { "type": "text" }
+}
