+++ frontend/package.json
@@ -19,21 +19,14 @@
 		"@formkit/auto-animate": "^1.0.0-pre-alpha.3",
 		"@playwright/test": "^1.38.1",
 		"@smui/button": "7.0.0-beta.15",
-		"@smui/card": "7.0.0-beta.15",
 		"@smui/checkbox": "7.0.0-beta.15",
 		"@smui/chips": "7.0.0-beta.15",
 		"@smui/circular-progress": "7.0.0-beta.15",
 		"@smui/data-table": "7.0.0-beta.15",
 		"@smui/dialog": "^7.0.0-beta.15",
-		"@smui/form-field": "7.0.0-beta.15",
 		"@smui/icon-button": "^7.0.0-beta.8",
-		"@smui/layout-grid": "7.0.0-beta.15",
-		"@smui/list": "7.0.0-beta.8",
 		"@smui/paper": "7.0.0-beta.14",
-		"@smui/ripple": "7.0.0-beta.15",
-		"@smui/segmented-button": "^7.0.0-beta.15",
 		"@smui/select": "7.0.0-beta.15",
-		"@smui/slider": "^7.0.0-beta.15",
 		"@smui/textfield": "^7.0.0-beta.15",
 		"@svelte-plugins/tooltips": "^0.1.6",
 		"@sveltejs/adapter-node": "^1.3.1",
@@ -72,6 +65,7 @@
 		"@auth/core": "^0.16.1",
 		"@auth/sveltekit": "^0.3.6",
 		"@mdi/js": "^7.3.67",
+		"@zeno-ml/zeno-instance-views": "^0.0.4",
 		"amazon-cognito-identity-js": "^6.3.6",
 		"d3-array": "^3.2.4",
 		"d3-interpolate": "^3.0.1",
