--- frontend/src/lib/components/instance-views/views/views/openai-chat/SystemBlock.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	export let input: string;
-</script>
-
-<p class="m-0">
-	<b>System:</b>
-	<span>{input}</span>
-</p>
