+++ frontend/src/lib/components/instances/ComparisonViewTableHeader.svelte
