+++ frontend/src/lib/components/popups/Popup.svelte
@@ -8,7 +8,7 @@
 	let paperHeight;
 </script>
 
-<div
+<button
 	class="absolute inset-0 z-20 flex justify-center items-baseline p-12 bg-grey bg-opacity-60"
 	transition:fade={{ duration: 200 }}
 	bind:clientHeight={paperHeight}
@@ -18,4 +18,4 @@
 	<Paper class="pt-3 flex flex-col" elevation={7} on:mousedown={(e) => e.stopPropagation()}>
 		<slot />
 	</Paper>
-</div>
+</button>
